+++ src/main/java/net/yxiao233/ifeu/IndustrialForegoingExtraUpgrades.java
@@ -18,6 +18,7 @@ import net.yxiao233.ifeu.common.networking.ModNetWorking;
 import net.yxiao233.ifeu.common.provider.ModRecipeProvider;
 import net.yxiao233.ifeu.common.provider.ModSerializableProvider;
 import net.yxiao233.ifeu.common.registry.*;
+import net.yxiao233.ifeu.common.structure.IFEUMultiBlockStructures;
 
 @Mod(IndustrialForegoingExtraUpgrades.MODID)
 public class IndustrialForegoingExtraUpgrades extends ModuleController{
@@ -26,6 +27,7 @@ public class IndustrialForegoingExtraUpgrades extends ModuleController{
     public IndustrialForegoingExtraUpgrades(){
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
         MinecraftForge.EVENT_BUS.register(this);
+        modEventBus.addListener(this::commonStep);
 
         ModContents.BLOCKS.register(modEventBus);
         ModContents.ITEMS.register(modEventBus);
@@ -40,6 +42,10 @@ public class IndustrialForegoingExtraUpgrades extends ModuleController{
             ItemBlockRenderTypes.setRenderLayer(ModBlocks.FLUID_CRAFTING_TABLE.getLeft().get(), RenderType.translucent());
             ItemBlockRenderTypes.setRenderLayer(ModBlocks.FLUID_TRANSFER.getLeft().get(), RenderType.translucent());
         }
+    }
+
+    private void commonStep(FMLCommonSetupEvent event){
+        IFEUMultiBlockStructures.init();
     }
 
     @Override